@frenzymadness 2021-09-15
Thanks Victor. I've incorporated the efficiency improvement proposed into the patch and I'll test it now. I will let you know about the results.