@frenzymadness 2021-09-15
Simple reproducer:
```python
from urllib2 import AbstractBasicAuthHandler

AbstractBasicAuthHandler.rx.search(
    "basic " + ("," * 25) + "A"
)
```
Official one using pyperf:
```python
from urllib2 import AbstractBasicAuthHandler
import pyperf


class AuthHandler(AbstractBasicAuthHandler):
    handler = None

    def retry_http_basic_auth(self, host, req, realm):
        self.realm = realm
        return None

realm = 'realm@example.com'
simple = 'Basic realm="{}"'.format(realm)
repeat_10 = '' + ',' * 10 + 'A'
repeat_10_2 = '' + ',' * (10 ** 2)
repeat_10_4 = '' + ',' * (10 ** 4)

class Headers:
    def __init__(self, header):
        self.header = header

    def get(self, *ignored_args):
        return self.header

    def get_all(self, *ignored_args):
        return [self.header]

def func(handler, headers):
    try:
        handler.http_error_auth_reqed("WWW-Authenticate", None, None, headers)
    except ValueError:
        pass


runner = pyperf.Runner()
handler = AuthHandler()

for name, header in (
    ('simple', simple),
    ('repeat 10', repeat_10),
    ('repeat 10^2', repeat_10_2),
    ('repeat 10^4', repeat_10_4),
):
    headers = Headers(header)
    runner.bench_func(name, func, handler, headers)
```

Old version:
```
# time python2 simple_redos.py 

real	0m45.528s
user	0m45.209s
sys	0m0.040s

# time python2 python2_redos.py
.....................
simple: Mean +- std dev: 2.82 us +- 0.19 us
.....................
repeat 10: Mean +- std dev: 427 us +- 27 us
^CBenchmark interrupted: exit

real	25m49.137s
user	25m35.931s
sys	0m1.769s
```
There is not a single dot for 10² in the benchmark above in more than 25 minutes.

New version:
```
# time python2 simple_redos.py 

real	0m0.039s
user	0m0.026s
sys	0m0.010s

# time python2 python2_redos.py 
.....................
simple: Mean +- std dev: 2.72 us +- 0.22 us
.....................
repeat 10: Mean +- std dev: 5.76 us +- 0.55 us
.....................
repeat 10^2: Mean +- std dev: 524 us +- 39 us
.....................
repeat 10^4: Mean +- std dev: 5.65 sec +- 0.48 sec

real	8m44.457s
user	8m36.923s
sys	0m2.129s
```
Benchmark is not as fast as it is for the new Pythons but it's able to finish testing in less than 9 minutes and provides reasonable results also for 10⁴.

I think this can be approved.