@vstinner 2021-09-15
```
repeat 10: Mean +- std dev: 5.76 us +- 0.55 us
repeat 10^2: Mean +- std dev: 524 us +- 39 us
repeat 10^4: Mean +- std dev: 5.65 sec +- 0.48 sec
```

LGTM.

https://bugs.python.org/issue39503 and https://bugs.python.org/issue43075 are about exponential complexity, something like: O(n^2). Here, it looks more like O(n) complexity which is the expected result.