@frenzymadness 2021-09-17
I've added one more patch here for CVE-2021-3737. The fix also contains a test and it passes. It is not possible to build Python 2.7 in rawhide now but I've managed to do that in mock with downgraded openssl to version 1.1 and here is verification of the fix using the official reproducer:

Old version:
```
[root@e5f1018c94e0 src]# python3 server.py &
[1] 72

[root@e5f1018c94e0 src]# time python2 client2.py 

… no response for minutes - infinite loop …

^CTraceback (most recent call last):
  File "client2.py", line 5, in <module>
    r = conn.getresponse()
  File "/usr/lib64/python2.7/httplib.py", line 1177, in getresponse
    response.begin()
  File "/usr/lib64/python2.7/httplib.py", line 457, in begin
    skip = self.fp.readline(_MAXLINE + 1)
  File "/usr/lib64/python2.7/socket.py", line 480, in readline
    data = self._sock.recv(self._rbufsize)
KeyboardInterrupt

real	3m2.678s
user	2m7.421s
sys	0m54.106s
```
New version:
```
[root@73a17da0838d src]# python3 server.py &
[1] 71

[root@73a17da0838d src]# python2 client2.py 
Traceback (most recent call last):
  File "client2.py", line 5, in <module>
    r = conn.getresponse()
  File "/usr/lib64/python2.7/httplib.py", line 1191, in getresponse
    response.begin()
  File "/usr/lib64/python2.7/httplib.py", line 475, in begin
    skipped_headers = _read_headers(self.fp)
  File "/usr/lib64/python2.7/httplib.py", line 381, in _read_headers
    raise HTTPException("got more than %d headers" % _MAXHEADERS)
httplib.HTTPException: got more than 100 headers
```

@vstinner could you please take a look once more? No rush, I guess we have to wait for openssl 1.1 packages in rawhide.