@vstinner 2021-09-20
> @vstinner could you please take a look once more? No rush, I guess we have to wait for openssl 1.1 packages in rawhide.

Hum, you may mention that it's a backport of: https://github.com/python/cpython/commit/47895e31b6f626bc6ce47d175fe9d43c1098909d

I checked the backport: yes, it LGTM ;-)