@hroncok 2021-09-20
For the record why I merged this: I've accidentally merged the Fedora PR, so I went ahead and merged this one as well.